refactor: consolidate OpampConfig and address review feedback
- Remove duplicate OpampConfig from ffwd-opamp, re-export from ffwd-config
- Add proper Default impl for OpampConfig with correct field defaults
- Remove unused `log` dependency from ffwd-opamp
- Document reload endpoint security (bind to localhost)
- Explain drain-before-validate design decision in bootstrap comment

Addresses review feedback: duplicate types, unused deps, security docs.
